pyproject.toml
[project]
name = "portainer-webhook"
version = "1.0.0"
description = "Simple webhook application to prompt updating stacks in Portainer"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "selenium>=4.40.0",
    "waitress>=3.0.2",
]
